requirements.txt
### base ###
django<1.6
pillow

### helpers ###
south
django-template-repl
django-admin-tools
django-filter
django-tables2
django-bootstrap3
django-admin-bootstrapped
pynliner
-e git://github.com/matuu/django-enhanced-cbv.git#egg=django-enhanced-cbv
